requirements.txt
easyocr==1.7.2
####
pytesseract==0.3.13
pandas==2.2.3
pillow
opencv-python
pyzbar